python_requirements.txt
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.35.0
transformers>=4.44.0
accelerate>=0.26.0
safetensors>=0.4.5
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
huggingface-hub>=0.20.0
sentencepiece>=0.1.99
protobuf>=4.21.0
tokenizers>=0.15.0
scipy>=1.10.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.8 